MCP/requirements.txt
# CryptoMind MCP Chatbot — Dependencies
# pip install -r requirements.txt

streamlit>=1.35.0
google-generativeai>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
